[IMP] l10n_hu_edi: Make timeout recovery more generic
Up to now, timeout recovery specifically applied to invoices in a
timeout state. However, the timeout recovery flow can apply to any
situation where an invoice's transaction code is unknown or has been
lost. This can be the case, for example, if the invoice was first
confirmed, then the transaction code was somehow lost. In such a case,
if the invoice is sent again, the NAV will return a duplicate invoice
error.

We rename 'timeout recovery' into 'transaction recovery' and make it
a generic flow which gets triggered every time the user uses the
'Update Status' button. This requests a list of all recent transactions
from the NAV, and attempts to match any unknown transaction IDs to
existing invoices in Odoo, updating the invoice state as appropriate.