[FIX] l10n_hu_edi: Set chain index on sending.
Instead of setting the chain index on posting, we set it on sending, and
un-set it if the invoice is rejected or cancelled.

When we send an invoice, we assign the first available chain index. This
ensures that when re-sending a rejected or cancelled invoice, the same
chain index is re-used.

In addition, we warn the user if they try to send an invoice but an
invoice that seems to be earlier in the chain is not sent.